# smart-ground-hub/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SmartGround API
  version: 2.0.0
  description: >
    REST API for IoT device management with MQTT-based device discovery and
    command signaling. Includes endpoints to manage device types, register
    devices, and trigger config pushes to SmartBox firmware via MQTT.

    ## Device Management

    All device operations use the flat `/api/devices` endpoint. Devices are
    always registered on a SmartBox (network access point) but are managed
    as a flat list with optional filtering by SmartBox or Range.
    Use **POST /api/devices** to register and immediately trigger a config push
    to the SmartBox firmware via MQTT.

    ## Breaking changes from v2.x

    - `POST /api/smart-boxes/{id}/devices` removed — use `POST /api/devices`
    - `GET /api/smart-boxes/{id}/devices` removed — use `GET /api/devices?smartBoxId=X`
    - `POST/DELETE /api/ranges/{rangeId}/devices/{deviceId}` removed — use `POST/DELETE /api/devices/{id}/range`
    - `SmartBoxResponse.deviceCount` removed
servers:
  - url: http://localhost:8080
    description: JVM App
  - url: http://localhost:8081
    description: Native App
security:
  - bearerAuth: []
paths:
...
  /api/device-types:
    get:
      summary: List all device types
      operationId: listDeviceTypes
      tags: [DeviceType]
      parameters:
        - name: firmwareConfigId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter device types by firmware configuration
      responses:
        '200':
          description: A list of device types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceTypeResponse'
  /api/device-types/{id}:
    get:
      summary: Get a device type by ID
      operationId: getDeviceType
      tags: [DeviceType]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device type found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeResponse'
        '404':
          description: Device type not found

    patch:
      summary: Update a device type
      operationId: updateDeviceType
      tags: [DeviceType]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceTypeRequest'
      responses:
        '200':
          description: Device type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeResponse'
        '404':
          description: Device type not found

  # ─────────────────────────────────────────────────────────────────────────────
  # DeviceTypeGroup (consolidated under /device-types)
  # ─────────────────────────────────────────────────────────────────────────────
...
  /api/device-types/groups/{id}:
    get:
      summary: Get a device type group by ID
      operationId: getDeviceTypeGroup
      tags: [DeviceType]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device type group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTypeGroupResponse'
        '404':
          description: Device type group not found

  # ─────────────────────────────────────────────────────────────────────────────
  # FirmwareConfig (consolidated under /device-types)
  # ─────────────────────────────────────────────────────────────────────────────
  /api/device-types/firmware-configs:
    get:
      summary: List all firmware configurations
      operationId: listFirmwareConfigs
      tags: [DeviceType]
      responses:
        '200':
          description: A list of firmware configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FirmwareConfigResponse'

    post:
      summary: Register a new firmware configuration
      description: >
        Creates a new firmware configuration with associated signal types and device
        type groups. Requires ADMIN role. Each entry in signal_types both defines a
        signal type and creates (or reuses) a named device type group.
      operationId: registerFirmwareConfig
      tags: [DeviceType]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterFirmwareConfigRequest'
      responses:
        '201':
          description: Firmware configuration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirmwareConfigResponse'
        '400':
          description: Invalid manifest or duplicate firmware version
        '403':
          description: User does not have ADMIN role
  /api/device-types/firmware-configs/{id}:
    get:
      summary: Get a firmware configuration by ID
      operationId: getFirmwareConfig
      tags: [DeviceType]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Firmware configuration found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirmwareConfigResponse'
        '404':
          description: Firmware configuration not found

  # ─────────────────────────────────────────────────────────────────────────────
  # Device
  # ─────────────────────────────────────────────────────────────────────────────
...
  /api/devices/{id}:
    get:
      summary: Get device by ID
      operationId: getDevice
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          description: Device not found

    patch:
      summary: Update device properties
      operationId: updateDevice
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceRequest'
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          description: Device not found

    delete:
      summary: Delete a device
      operationId: deleteDevice
      tags: [Device]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Device deleted
        '404':
          description: Device not found
...
components:
  schemas:
...
    DeviceTypeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Human-readable label (e.g. "Wurfmaschine Typ A")
        signalTypeId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        groupName:
          type: string
          description: Name of the device type group this type belongs to
        signalDurationMs:
          type: integer
          description: How long the signal is held active (ms)
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before the signal fires (ms)
        command:
          type: string
          description: Command string sent to the firmware (e.g. ON, OFF)
        device:
          type: string
          description: Device kind as understood by the firmware (e.g. SENSOR, ACTUATOR)
        direction:
          type: string
          description: Communication direction (INPUT or OUTPUT)
    DeviceTypeGroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Group name (e.g. "Wurfmaschine", "LED-Anzeige")
    FirmwareConfigResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: string
          description: Firmware version string (e.g. "0.6v")
        boxType:
          type: string
          description: Hardware variant this firmware targets
        signalTypes:
          type: array
          items:
            $ref: '#/components/schemas/FirmwareConfigSignalTypeResponse'
    FirmwareConfigSignalTypeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        communicationDirection:
          type: string
          description: INPUT or OUTPUT
        device:
          type: string
          description: SENSOR or ACTUATOR
        command:
          type: string
          description: Command string sent to firmware
    RegisterFirmwareConfigRequest:
      type: object
      required: [version, box_type, signal_types]
      properties:
        version:
          type: string
          description: Firmware version string (must be unique per box_type)
        box_type:
          type: string
          description: Hardware variant identifier
        signal_types:
          type: array
          items:
            $ref: '#/components/schemas/RegisterFirmwareConfigSignalTypeEntry'
    RegisterFirmwareConfigSignalTypeEntry:
      type: object
      required: [direction, device, command, group_name, name, signal_duration_ms]
      properties:
        direction:
          type: string
          description: CommunicationDirection enum value (e.g. OUTPUT)
        device:
          type: string
          description: DeviceKind enum value (e.g. ACTUATOR)
        command:
          type: string
          description: Firmware command string (e.g. FIRE)
        group_name:
          type: string
          description: Name of the device type group to create or reuse
        name:
          type: string
          description: Name for the device type created from this signal type
        signal_duration_ms:
          type: integer
          minimum: 0
          description: Signal duration in milliseconds
        delay_signal_duration_ms:
          type: integer
          nullable: true
          description: Optional delay before the signal in milliseconds
    UpdateDeviceTypeRequest:
      type: object
      properties:
        name:
          type: string
          description: User-friendly label for this device type (e.g. "Wurfmaschine Typ A")
        signalDurationMs:
          type: integer
          minimum: 0
          nullable: true
          description: How long the signal is held active (ms)
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before the signal fires (ms)

    # ───── Device ─────────────────────────────────────────────────────────────
    DeviceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alias:
          type: string
        smartBoxId:
          type: string
          format: uuid
          description: ID of the SmartBox this device is connected to
        groupId:
          type: string
          format: uuid
          description: Device type group
        groupName:
          type: string
          description: Name of the device type group
        deviceTypeId:
          type: string
          format: uuid
        deviceType:
          type: string
        command:
          type: string
          description: Firmware command (e.g. FIRE, ON)
        boxType:
          type: string
        blocked:
          type: boolean
        healthy:
          type: boolean
        signalDurationMs:
          type: integer
          description: How long the signal is held active
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Optional delay before signal fires
        pinConfig:
          $ref: '#/components/schemas/PinConfig'
        rangeId:
          type: string
          format: uuid
          nullable: true
...
    PinConfig:
      type: object
      required: [pin, mode]
      properties:
        pin:
          type: integer
          minimum: 0
          maximum: 40
          description: GPIO pin number on the Raspberry Pi Pico 2W (valid range 0–29 for user GPIOs; 0–40 reserved)
        mode:
          type: string
          enum: [OUTPUT, INPUT, INPUT_PULLUP, INPUT_PULLDOWN, PWM]
          description: GPIO mode as understood by the firmware
        pwmFrequencyHz:
          type: integer
          minimum: 1
          description: Required when mode is PWM. Sets the PWM carrier frequency.
        invertLogic:
          type: boolean
          default: false
          description: When true, the firmware inverts the signal (active-low wiring)

    # ───── Pagination ──────────────────────────────────────────────────────────
...
    UpdateDeviceRequest:
      type: object
      properties:
        alias:
          type: string
        signalDurationMs:
          type: integer
          nullable: true
          description: How long the signal is held active (null = use device type default)
        delaySignalDurationMs:
          type: integer
          nullable: true
          description: Delay before signal fires (null = use device type default)
        groupId:
          type: string
          format: uuid
          description: Reassign device to a different device type group
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT